Closes #50: Configuration two resource servers (local authorization server + new ixortalk-authorization-server module)
