+++ pom.xml
@@ -36,7 +36,7 @@
 
     <groupId>com.ixortalk</groupId>
     <artifactId>ixortalk-authserver</artifactId>
-    <version>1.2.4-SNAPSHOT</version>
+    <version>1.3.0-SNAPSHOT</version>
     <packaging>war</packaging>
     <name>ixortalk-authserver</name>
     <description>IxorTalk Auth Server (uaa)</description>
